requirements.txt
tensorflow>=2.9.0
keras>=2.9.0
numpy>=1.22.0
matplotlib>=3.5.0
pillow>=9.0.0
scikit-learn>=1.0.0
h5py>=3.6.0
ipykernel>=6.0.0
jupyter>=1.0.0
pandas>=1.3.0
